requirements.txt
discord==2.1.0
dpytest==0.6.3
python-decouple==3.6
pytest-asyncio==0.20.3
